+++ Code/Genrpg.Shared/PlayMultiplier/Settings/PlayMultSettings.cs
@@ -9,7 +9,7 @@ namespace Genrpg.Shared.PlayMultiplier.Settings
     {
         public override string Id { get; set; }
         public double ExtraDailyDistPerTotalDice { get; set; }
-        public long MaxPlayMult { get; set; }
+        public int MaxPlayMult { get; set; }
         public double MaxMultAsPercentOfCurrentDice { get; set; }
     }
