+++ Code/Genrpg.Shared/Trader/Buffs/Interfaces/ITravelBuff.cs
@@ -2,9 +2,9 @@
 {
     public interface ITravelBuff
     {
-        long BonusSpeed { get; set; }
-        long Capacity { get; set; }
-        long RationsCost { get; set; }
+        int BonusSpeed { get; set; }
+        int Capacity { get; set; }
+        int RationsCost { get; set; }
         double ForageChance { get; set; }
         double GoodEventChance { get; set; }
         double BadEventChance { get; set; }
