+++ Code/Genrpg.Shared/Trader/Caravans/Entities/CaravanStatusBlock.cs
@@ -2,13 +2,13 @@
 {
     public class CaravanStatusBlock
     {
-        public long DiceSpeed { get; set; }
-        public long ItemCount { get; set; }
+        public int DiceSpeed { get; set; }
+        public int ItemCount { get; set; }
 
         // Add bonuses for movement stats here.
-        public long BonusSpeed { get; set; }
-        public long RationsCost { get; set; }
-        public long Capacity { get; set; }
+        public int BonusSpeed { get; set; }
+        public int RationsCost { get; set; }
+        public int Capacity { get; set; }
 
         // Add chances for certain events here
         public double ForageChance { get; set; }
