+++ Code/Genrpg.Shared/Trader/Travel/Entities/TravelDay.cs
@@ -5,15 +5,15 @@ namespace Genrpg.Shared.Trader.Travel.Entities
 {
     public class TravelDay
     {
-        public List<long> RolledDistances { get; set; } = new List<long>();
-        public long BonusDistance { get; set; }
-        public long TotalDistance { get; set; }
-        public long EndDistance { get; set; }
-        public long Day { get; set; }
-        public long EndFlags { get; set; }
-        public long EndDiceSpeed { get; set; }
-        public long EndBonusSpeed { get; set; }
-        public long RationsCost { get; set; }
+        public List<int> RolledDistances { get; set; } = new List<int>();
+        public int BonusDistance { get; set; }
+        public int TotalDistance { get; set; }
+        public int EndDistance { get; set; }
+        public int Day { get; set; }
+        public int EndFlags { get; set; }
+        public int EndDiceSpeed { get; set; }
+        public int EndBonusSpeed { get; set; }
+        public int RationsCost { get; set; }
         public List<Reward> TravelRewards { get; set; } = new List<Reward>();
     }
 }
