+++ Code/Genrpg.Shared/Trader/Travel/WebApi/HeadToTargetRequest.cs
@@ -4,8 +4,8 @@ namespace Genrpg.Shared.Trader.Travel.WebApi
 {
     public class HeadToTargetRequest : IClientUserRequest
     {
-        public long ToX { get; set; }
-        public long ToY { get; set; }
-        public long ToCityId { get; set; }
+        public int ToX { get; set; }
+        public int ToY { get; set; }
+        public int ToCityId { get; set; }
     }
 }
