+++ Code/Genrpg.Shared/Trader/Travel/WebApi/TravelResponse.cs
@@ -8,16 +8,16 @@ namespace Genrpg.Shared.Trader.Travel.WebApi
     {
         public bool Success { get; set; }
         public string ErrorMessage { get; set; }
-        public long TargetCityId { get; set; }
-        public long TotalDistanceTravelled { get; set; }
-        public long DistanceLeft { get; set; }
-        public long TotalCost { get; set; }
-        public long DistanceAlongRoad { get; set; }
-        public long EndDay { get; set; }
-        public long EnterCityId { get; set; }
+        public int TargetCityId { get; set; }
+        public int TotalDistanceTravelled { get; set; }
+        public int DistanceLeft { get; set; }
+        public int TotalCost { get; set; }
+        public int DistanceAlongRoad { get; set; }
+        public int EndDay { get; set; }
+        public int EnterCityId { get; set; }
         public List<TravelDay> Days { get; set; } = new List<TravelDay>();
         public List<string> Messages { get; set; } = new List<string>();
-        public long EndCoreFlags { get; set; }
+        public int EndCoreFlags { get; set; }
     }
 
 }
